Añadiendo clases

+++ src/main/java/pe/edu/idat/EF_FloresRoman/Dto/InventarioDto.java
@@ -1,18 +1,12 @@
 package pe.edu.idat.EF_FloresRoman.Dto;
 public class InventarioDto {
-    private Integer idProduct;
+    private Long idProduct;
     private Integer cantidadReal;
     // Getters and setters
-    public Integer getIdProduct() {
+    public Long getIdProduct() {
         return idProduct;
     }
-    public void setIdProduct(Integer idProduct) {
-        this.idProduct = idProduct;
-    }
-    public Integer getCantidadReal() {
-        return cantidadReal;
-    }
-    public void setCantidadReal(Integer cantidadReal) {
-        this.cantidadReal = cantidadReal;
-    }
+    public void setIdProduct(Long idProduct) {this.idProduct = idProduct;}
+    public Integer getCantidadReal() {return cantidadReal;}
+    public void setCantidadReal(Integer cantidadReal) {this.cantidadReal = cantidadReal;}
 }
